% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_tidy.R
\name{eav_to_tidy}
\alias{eav_to_tidy}
\title{assumes `value` and `variable` columns
eav_to_tidy(eav_tiny)}
\usage{
eav_to_tidy(data, id_cols = NULL)
}
\description{
assumes `value` and `variable` columns
eav_to_tidy(eav_tiny)
}
