% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_eav}
\alias{filter_eav}
\title{Filter an EAV tibble}
\usage{
filter_eav(.data, ...)
}
\arguments{
\item{.data}{a tbl_eav}

\item{...}{filter conditions}

\item{.c}{a single condition (to be replaced by dots...)}
}
\description{
Filter an EAV tibble
}
\examples{
eav_tiny \%>\% group_by(empid) \%>\% filter_eav(lu("first") == "John")

eav_tiny \%>\% filter_eav(is.na(last))

# only gets explicit NA
tinyeav \%>\%
  filter(var == "last", is.na(value)) \%>\% {
  semi_join(tinyeav, ., by = "empid")
  }

# reversing the condition specially for NA
tinyeav \%>\%
  filter(var == "last", !is.na(value)) \%>\% {
  anti_join(tinyeav, ., by = "empid")}

# or first make missing data explicit
tinyeav \%>\%
  complete(nesting(empid), nesting(var)) \%>\%
  filter(var == "last", is.na(value)) \%>\% {
  semi_join(tinyeav, ., by = "empid")}

# or tidy first (this version makes it explicit)
tinyeav \%>\%
  spread(var, value) \%>\%
  filter(is.na(last)) \%>\%
  gather("var", "value", -empid) \%>\%
  arrange(empid)

# multiple conditions
tinyeav \%>\%
  filter_eav(first == "John" & last == "Smith")

tinyeav \%>\%
  group_by(empid) \%>\%
  summarize(include_row =
  						any(var == "first" & value == "John") &
  						any(var == "last" & value == "Smith")) \%>\%
  filter(include_row) \%>\% {
  semi_join(tinyeav, ., by = "empid")}



}
