% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%forwhich\%}
\alias{\%forwhich\%}
\title{Return a single value from a vector, for which a different condition holds}
\usage{
x \%forwhich\% p
}
\arguments{
\item{x}{a vector from which to return results}

\item{p}{a vector of booleans (or other `[` indices?) length of x}
}
\description{
Intended to help with EAV calculations that potentially have length 0, ruining data frame operations.
}
