% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lu}
\alias{lu}
\title{Lookup a an EAV value-variable pair when inside a tidyeval context}
\usage{
lu(varname)
}
\description{
Lookup a an EAV value-variable pair when inside a tidyeval context
}
\examples{
eav_tiny \%>\% group_by(empid) \%>\% mutate(last = lu("last"))
eav_tiny \%>\% group_by(empid) \%>\% filter_eav(lu("last") == "Smith")
}
