% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate_eav}
\alias{mutate_eav}
\title{Mutate an EAV table (e.g., add/modify rows)}
\usage{
mutate_eav(.data, ...)
}
\arguments{
\item{...}{}
}
\description{
Mutate an EAV table (e.g., add/modify rows)
}
\examples{
tinyeav \%>\%
  mutate_eav(middle = "Ramsey")

tinyeav \%>\%
  group_by(empid) \%>\%
  summarize(var = "middle",
            value = "Ramsey") \%>\% {
    bind_rows(tinyeav, .)}

tinyeav \%>\%
  mutate_eav(middle = if_else(first == "John", "Ramsey", middle))

tinyeav \%>\%
  group_by(empid) \%>\%
  summarize(middle = if_else(any(var == "first" & value == "John"), "Ramsey", lookup(value, var == "middle") ))

}
