% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%0\%}
\title{Helper to replace empty lists}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{vctrs}{\code{\link[vctrs]{\%0\%}}}
}}

