% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_eav.R
\name{tbl_eav}
\alias{tbl_eav}
\title{Make an EAV tibble out of a regular tibble}
\usage{
tbl_eav(.data, id_cols = NULL, variable_col = c("variable", "var"),
  value_col = "value")
}
\arguments{
\item{.data}{starting tibble in EAV format}

\item{id_cols}{specification of columns which represent entities}

\item{variable_col}{spec of column which represent variable names}

\item{value_col}{spec of column which represent values

Columns may be specified using character vectors or (hopefully) tidyselect semantics.}
}
\description{
Make an EAV tibble out of a regular tibble
}
