% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_tidy.R
\name{tidy_to_eav}
\alias{tidy_to_eav}
\title{tidy_to_eav(eav_to_tidy(eav_tiny), eav_tiny)}
\usage{
tidy_to_eav(data, orig = NULL, id_cols = NULL)
}
\description{
tidy_to_eav(eav_to_tidy(eav_tiny), eav_tiny)
}
