% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_eav.R
\name{tbl_eav}
\alias{tbl_eav}
\title{Make an EAV tibble out of a regular tibble}
\usage{
tbl_eav(.data, entity_cols, attr_cols, val_cols)
}
\arguments{
\item{.data}{starting tibble in EAV format}

\item{entity_cols}{specification of columns which represent entities}

\item{attr_cols}{spec of columns which represent attributes}

\item{val_cols}{spec of columns which represent values

Columns may be specified using character vectors or tidyselect semantics.}
}
\description{
Make an EAV tibble out of a regular tibble
}
